refactor(GamePiece): 게임을 진행하는 기물들을 묶는 추상 클래스 생성(Blank 제외)
