+++ src/main/java/chess/domain/piece/Blank.java
@@ -6,7 +6,7 @@ import chess.exception.InvalidMethodCallException;
 import chess.exception.PieceDoesNotExistException;
 import java.util.List;
 
-public class Blank extends Piece {
+public final class Blank extends Piece {
 
     private static final String INITIAL = ".";
     private static final Blank CACHED_BLANK = new Blank();
@@ -17,6 +17,11 @@ public class Blank extends Piece {
 
     public static Blank getBlank() {
         return CACHED_BLANK;
+    }
+
+    @Override
+    public List<Position> route(Position from, Position to) {
+        throw new PieceDoesNotExistException();
     }
 
     @Override
