+++ src/main/java/chess/domain/piece/Piece.java
@@ -2,10 +2,8 @@ package chess.domain.piece;
 
 import chess.domain.Side;
 import chess.domain.position.Position;
-import chess.exception.InvalidMovementException;
 import java.util.List;
 import java.util.Locale;
-import java.util.Objects;
 
 public abstract class Piece {
 
@@ -18,20 +16,7 @@ public abstract class Piece {
         this.initial = initial;
     }
 
-    public List<Position> route(Position from, Position to) {
-        if (Objects.equals(from, to)) {
-            throw new InvalidMovementException("자기 자신의 위치로는 이동할 수 없습니다.");
-        }
-
-        int rowDifference = Position.differenceOfRow(from, to);
-        int columnDifference = Position.differenceOfColumn(from, to);
-
-        if (movable(rowDifference, columnDifference)) {
-            return getRoute(from, to);
-        }
-
-        throw new InvalidMovementException("해당 기물의 이동 룰에 어긋납니다.");
-    }
+    public abstract List<Position> route(Position from, Position to);
 
     protected abstract boolean movable(int rowDifference, int columnDifference);
 
@@ -57,7 +42,11 @@ public abstract class Piece {
         return this.side == side;
     }
 
-    public String getInitial() {
+    public abstract boolean diagonal(Position from, Position to);
+
+    public abstract boolean forward(Position from, Position to);
+
+    public final String getInitial() {
         if (side == Side.WHITE) {
             return initial.toLowerCase(Locale.ROOT);
         }
@@ -66,8 +55,4 @@ public abstract class Piece {
         }
         return initial;
     }
-
-    public abstract boolean diagonal(Position from, Position to);
-
-    public abstract boolean forward(Position from, Position to);
 }
